[logging] v0.12.12, 29 dec 2007 - added destination::named - similar to formatter::named_spacer, but for destinations
SVN-Revision: 42337

+++ logging/boost/logging/format/formatter/high_precision_time.hpp
@@ -77,7 +77,7 @@ template<class convert = do_convert_format::prepend> struct high_precision_time_
     template<class msg_type> void write_high_precision_time(msg_type & msg, ::boost::posix_time::ptime val) const {
         char_type buffer[64];
 
-        int nanosecs = val.time_of_day().fractional_seconds();
+        int nanosecs = (int)val.time_of_day().fractional_seconds();
         int digits = val.time_of_day().num_fractional_digits();
         switch ( digits) {
         case 0: break; // no high precision at all
